api: move tests to .old package
